requirements.txt
google-cloud-bigquery==1.5.0
